backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
fastapi-users[sqlalchemy]>=13.0.0
httpx>=0.27.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
alembic>=1.13.0
faiss-cpu>=1.8.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
argon2-cffi>=25.1.0
